requirements.txt
asgiref==3.8.1
Django==4.2.2
djangorestframework==3.15.2
Faker==30.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
